# app/backend/api-gateway/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Task Tracker API
  version: "1.0.0"
  description: |
    External REST API exposed by the api-gateway.

    **Internal** service-to-service communication (api-gateway → auth-service,
    api-gateway → project-service) uses gRPC and is defined separately in
    `auth-service/api/proto/auth_service.proto` and
    `project-service/api/proto/project_service.proto`.
    Those contracts are intentionally NOT duplicated here.
servers:
  - url: http://localhost:8000
    description: Local development (via api-gateway)

# ── Security ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:          { type: integer }
        project_id:  { type: integer }
        assignee_id: { type: integer, nullable: true }
        name:        { type: string }
        description: { type: string, nullable: true }
        priority:    { type: string, enum: [TASK_PRIORITY_LOW, TASK_PRIORITY_MEDIUM, TASK_PRIORITY_HIGH] }
        difficulty:  { type: string, enum: [TASK_DIFFICULTY_EASY, TASK_DIFFICULTY_MEDIUM, TASK_DIFFICULTY_HARD] }
        status:      { type: string, enum: [TASK_STATUS_NOT_STARTED, TASK_STATUS_IN_WORK, TASK_STATUS_ON_REVIEW, TASK_STATUS_COMPLETED] }
        start_date:  { type: string, format: date-time, nullable: true }
        end_date:    { type: string, format: date-time, nullable: true }
...
    OperationResponse:
      type: object
      properties:
        message: { type: string }

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /api/projects:
    post:
      tags: [Projects]
      summary: Create a new project
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description]
              properties:
                name:        { type: string }
                description: { type: string }
      responses:
        "201":
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
  /api/projects/{id}/tasks:
    get:
      tags: [Projects]
      summary: Get tasks for a project (members only)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks: { type: array, items: { $ref: "#/components/schemas/Task" } }
...
  /api/tasks:
    post:
      tags: [Tasks]
      summary: Create a task (manager only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, name, priority, difficulty]
              properties:
                project_id:  { type: integer }
                name:        { type: string }
                description: { type: string }
                priority:    { type: integer, description: "1=LOW 2=MEDIUM 3=HIGH" }
                difficulty:  { type: integer, description: "1=EASY 2=MEDIUM 3=HARD" }
      responses:
        "201": { description: Task created }
...
  /api/admin/projects:
    put:
      tags: [Admin]
      summary: Update a project
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationResponse" }
    delete:
      tags: [Admin]
      summary: Delete a project
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationResponse" }
...